introduce quay_base
