+++ reconcile/quay_permissions.py
@@ -1,10 +1,8 @@
 import logging
 
 import utils.gql as gql
-from utils.secret_reader import SecretReader
-import reconcile.queries as queries
 
-from utils.quay_api import QuayApi
+from reconcile.quay_base import get_quay_api_store
 
 
 QUAY_REPOS_QUERY = """
@@ -38,20 +36,6 @@ QUAY_REPOS_QUERY = """
 """
 
 QONTRACT_INTEGRATION = 'quay-permissions'
-
-
-def get_quay_api_store():
-    quay_orgs = queries.get_quay_orgs()
-    settings = queries.get_app_interface_settings()
-    secret_reader = SecretReader(settings=settings)
-    store = {}
-    for org_data in quay_orgs:
-        name = org_data['name']
-        server_url = org_data.get('serverUrl')
-        token = secret_reader.read(org_data['automationToken'])
-        store[name] = QuayApi(token, name, base_url=server_url)
-
-    return store
 
 
 def run(dry_run):
